refactor: ServiceとUseCaseの重複を解消し、共通utilを整理
- service/ディレクトリを削除（UseCaseに統合）
- service/utilをshared/utilに移動
- import文を新しいパス構造に更新
- Clean Architecture準拠へのリファクタリング
