--- backend/main/service/util/error_util.go
@@ -1,7 +1,0 @@
-package util
-
-import "errors"
-
-var ErrUnauthorized = errors.New("unauthorized")
-
-var ErrForbidden = errors.New("forbidden")
