+++ backend/main/shared/util/error_util.go
@@ -1,0 +1,7 @@
+package util
+
+import "errors"
+
+var ErrUnauthorized = errors.New("unauthorized")
+
+var ErrForbidden = errors.New("forbidden")
